fix: fix storageFrom.jsx rename failure
