+++ src/pages/cluster/components/plugin/index.less
