+++ src/pages/cluster/components/plugin/uiSchemas.js
