+++ src/pages/cluster/containers/Template/cluster/actions/Create/Storage.jsx
@@ -19,7 +19,7 @@ import Base from 'components/Form';
 import { rootStore } from 'stores';
 import { toJS } from 'mobx';
 import { filter } from 'lodash';
-import StorageForm from 'pages/cluster/components/PluginForm/StorageFrom';
+import StorageForm from 'pages/cluster/components/plugin/StorageForm';
 
 @observer
 export default class Storage extends Base {
